requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
cryptography==42.0.2
pyinstaller==6.12.0
Pillow==10.2.0
selenium==4.18.1
webdriver-manager==4.0.1